+++ frontend/src/shared/Text.js
@@ -3,10 +3,15 @@
 import React from "react";
 import "./Text.css";
 
-const DisplayHeading = ({ title, subtitle }) => (
-	<h1 className="text-center mb-3 display-heading">
-		{title} <br /> <div className="green merriweather-black">{subtitle}</div>
-	</h1>
+const DisplayHeading = ({ title, subtitle, isLeftAlignment = false }) => (
+  <h1
+    className={`mb-3 display-heading ${
+      isLeftAlignment ? `text-left` : "text-center"
+    }`}
+  >
+    {title} <br /> <div className="green merriweather-black">{subtitle}</div>{" "}
+  </h1>
 );
+DisplayHeading.defaultProps = { alignment: false };
 
 export default DisplayHeading;
